+++ projects/ngx-testing-tools/src/lib/component/action/index.ts
@@ -1,2 +1,2 @@
-export { click } from './click';
-export { emitOutput } from './output';
+export { click } from './utils/click';
+export { emitOutput } from './utils/output';
