+++ projects/ngx-testing-tools/src/lib/component/action/utils/click.ts
