+++ projects/ngx-testing-tools/src/lib/component/query/index.ts
@@ -1,8 +1,8 @@
-export { findComponent } from './find-component';
-export { findAllComponents } from './find-all-components';
+export { findComponent } from './utils/find-component';
+export { findAllComponents } from './utils/find-all-components';
 
-export { findElement } from './find-element';
-export { findAllElements } from './find-all-elements';
+export { findElement } from './utils/find-element';
+export { findAllElements } from './utils/find-all-elements';
 
-export { findDebugElement } from './find-debug-element';
-export { findAllDebugElements } from './find-all-debug-elements';
+export { findDebugElement } from './utils/find-debug-element';
+export { findAllDebugElements } from './utils/find-all-debug-elements';
