+++ projects/ngx-testing-tools/src/lib/component/query/utils/find-component.ts
