+++ projects/ngx-testing-tools/src/lib/component/test-bed/component-tools.ts
@@ -1,12 +1,12 @@
 import { ComponentFixture } from '@angular/core/testing';
 import { assertInstance } from '../../common/assertion/assert-instance';
 import { buildBaseTools } from '../../common/test-bed/base/base-tools';
-import { buildComponentActionTools } from './component-action-tools';
-import { buildComponentQueryTools } from './component-query-tools';
+import { buildComponentActionTools } from '../action/component-action-tools';
+import { ComponentActionTools } from '../action/models/component-action-tools.model';
+import { buildComponentQueryTools } from '../query/component-query-tools';
+import { ComponentQueryTools } from '../query/models/component-query-tools.model';
 import { ComponentTestBedFactory } from './component-test-bed-factory';
 import { ComponentTools } from './models';
-import { ComponentActionTools } from './models/component-action-tools.model';
-import { ComponentQueryTools } from './models/component-query-tools.model';
 
 export function buildComponentTools<T>(factory: ComponentTestBedFactory<T>): ComponentTools<T> {
   const fixture: ComponentFixture<T> = factory['fixture'];
