+++ projects/ngx-testing-tools/src/lib/component/test-bed/models/component-tools.model.ts
@@ -1,8 +1,8 @@
 import { DebugElement } from '@angular/core';
 import { ComponentFixture } from '@angular/core/testing';
 import { BaseTools } from '../../../common/test-bed/base/models/base-tools.model';
-import { ComponentActionTools } from './component-action-tools.model';
-import { ComponentQueryTools } from './component-query-tools.model';
+import { ComponentActionTools } from '../../action/models/component-action-tools.model';
+import { ComponentQueryTools } from '../../query/models/component-query-tools.model';
 
 export interface ComponentTools<T, I extends {} = {}> extends BaseTools<I> {
   /**
