+++ projects/ngx-testing-tools/src/tests/integrations/components/action/event.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture } from '@angular/core/testing';
-import { emitEvent } from '../../../../lib/component/action/event';
+import { emitEvent } from '../../../../lib/component/action/utils/event';
 import { OuterComponent } from '../../../fixtures/components/outer.component';
 import { MyButtonDirective } from '../../../fixtures/directives/my-button.directive';
 import { createOuterComponentFixture } from '../../../fixtures/helpers/configuration/create-outer-component-fixture';
