+++ projects/ngx-testing-tools/src/tests/units/components/query/cannot-find.spec.ts
@@ -1,4 +1,4 @@
-import { throwCannotFind } from '../../../../lib/component/query/cannot-find';
+import { throwCannotFind } from '../../../../lib/component/query/utils/cannot-find';
 import { NoWhereDirective } from '../../../fixtures/directives/no-where.directive';
 
 describe('throwCannotFind', () => {
